refactor: removed external event handlers and macroquad demo
Only `draw` method from event handler actually used context, and it is unlikely to get used directly without yakui-miniquad getting wrapped in a custom event handler (like in the example)
